+++ src/game/GoodsEntry.java
@@ -4,12 +4,12 @@ public class GoodsEntry
 {
 
 	int waiting_acceptance;
-	byte days_since_pickup;
-	byte rating;
+	int days_since_pickup;
+	int rating;
 	int enroute_from;
 	int enroute_time;
-	byte last_speed;
-	byte last_age;
+	int last_speed;
+	int last_age;
 	int feeder_profit;
